[refactor] Members 패키지 domain 패키지 하위로 옮김
