+++ src/main/java/com/be/rebook/domain/members/jwt/CustomLogoutFilter.java
@@ -1,6 +1,6 @@
-package com.be.rebook.members.jwt;
+package com.be.rebook.domain.members.jwt;
 
-import com.be.rebook.members.repository.RefreshTokensRepository;
+import com.be.rebook.domain.members.repository.RefreshTokensRepository;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
