+++ src/main/java/com/be/rebook/domain/members/jwt/JWTFilter.java
@@ -1,7 +1,7 @@
-package com.be.rebook.members.jwt;
+package com.be.rebook.domain.members.jwt;
 
-import com.be.rebook.members.dto.CustomUserDetails;
-import com.be.rebook.members.entity.Members;
+import com.be.rebook.domain.members.entity.Members;
+import com.be.rebook.domain.members.dto.CustomUserDetails;
 import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
