+++ src/main/java/com/be/rebook/domain/members/repository/MembersRepository.java
@@ -1,6 +1,6 @@
-package com.be.rebook.members.repository;
+package com.be.rebook.domain.members.repository;
 
-import com.be.rebook.members.entity.Members;
+import com.be.rebook.domain.members.entity.Members;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MembersRepository extends JpaRepository<Members, Integer> {
