+++ src/main/java/com/be/rebook/domain/members/service/CustomUserDetailService.java
@@ -1,8 +1,8 @@
-package com.be.rebook.members.service;
+package com.be.rebook.domain.members.service;
 
-import com.be.rebook.members.dto.CustomUserDetails;
-import com.be.rebook.members.entity.Members;
-import com.be.rebook.members.repository.MembersRepository;
+import com.be.rebook.domain.members.dto.CustomUserDetails;
+import com.be.rebook.domain.members.entity.Members;
+import com.be.rebook.domain.members.repository.MembersRepository;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
